+++ apps/angular-example/src/app/test_components/c.component.ts
@@ -1,8 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'c-component',
-  template: '<div>TEST C</div>',
-  styles: [''],
-})
-export class CTestComponent {}