+++ apps/angular-example/src/app/test_components/d.component.ts
@@ -1,8 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'd-component',
-  template: '<div>TEST D</div>',
-  styles: [''],
-})
-export class DTestComponent {}