+++ apps/angular-example/src/app/test_components/flag.component.ts
@@ -1,8 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'flag-component',
-  template: '<div>this is a cool feature</div>',
-  styles: [''],
-})
-export class FlagComponent {}