+++ packages/node/src/express/index.ts
@@ -1,18 +1,17 @@
 import type { NextFunction, Request, Response } from "express";
 import { ABConfig, Abby, AbbyConfig, RemoteConfigValueString } from "@tryabby/core";
-import { F } from "ts-toolbelt";
 import { createAbby } from "../index";
 
 const instanceMap = new Map<string, Abby<any, any, any, any, any>>();
 const initializedProjects = new Set<string>();
 
 export function createAbbyMiddleWare<
-  FlagName extends string,
-  TestName extends string,
-  Tests extends Record<TestName, ABConfig>,
-  RemoteConfig extends Record<RemoteConfigName, RemoteConfigValueString>,
-  RemoteConfigName extends Extract<keyof RemoteConfig, string>,
->(config: F.Narrow<AbbyConfig<FlagName, Tests, string[], RemoteConfigName, RemoteConfig>>) {
+  const FlagName extends string,
+  const TestName extends string,
+  const Tests extends Record<TestName, ABConfig>,
+  const RemoteConfig extends Record<RemoteConfigName, RemoteConfigValueString>,
+  const RemoteConfigName extends Extract<keyof RemoteConfig, string>,
+>(config: AbbyConfig<FlagName, Tests, string[], RemoteConfigName, RemoteConfig>) {
   let abbyInstance = instanceMap.get(config.projectId);
 
   if (!abbyInstance) {